Move profile into users context
