+++ lib/chat_api_web/channels/conversation_channel.ex
@@ -2,7 +2,7 @@ defmodule ChatApiWeb.ConversationChannel do
   use ChatApiWeb, :channel
 
   alias ChatApiWeb.Presence
-  alias ChatApi.{Accounts, Messages, Conversations, Emails}
+  alias ChatApi.{Messages, Conversations, Emails}
 
   @impl true
   def join("conversation:lobby", payload, socket) do
