+++ lib/chat_api_web/controllers/user_profile_controller.ex
@@ -1,13 +1,13 @@
 defmodule ChatApiWeb.UserProfileController do
   use ChatApiWeb, :controller
 
-  alias ChatApi.UserProfiles
+  alias ChatApi.Users
 
   action_fallback ChatApiWeb.FallbackController
 
   def show(conn, _params) do
     with %{id: user_id} <- conn.assigns.current_user do
-      user_profile = UserProfiles.get_user_profile(user_id)
+      user_profile = Users.get_user_profile(user_id)
       render(conn, "show.json", user_profile: user_profile)
     end
   end
@@ -15,7 +15,7 @@ defmodule ChatApiWeb.UserProfileController do
   def create_or_update(conn, %{"user_profile" => user_profile_params}) do
     with %{id: user_id} <- conn.assigns.current_user do
       params = Map.merge(user_profile_params, %{"user_id" => user_id})
-      {:ok, user_profile} = UserProfiles.create_or_update_profile(user_id, params)
+      {:ok, user_profile} = Users.create_or_update_profile(user_id, params)
 
       render(conn, "show.json", user_profile: user_profile)
     end
